chore: move workforce under society & small doc polish
