+++ README.md
@@ -116,7 +116,7 @@ conda create --name camel python=3.10
 conda activate camel
 
 # Clone github repo
-git clone -b v0.2.4 https://github.com/camel-ai/camel.git
+git clone -b v0.2.5 https://github.com/camel-ai/camel.git
 
 # Change directory into project directory
 cd camel
