+++ camel/societies/workforce/__init__.py
