+++ camel/societies/workforce/prompts.py
