+++ camel/societies/workforce/role_playing_worker.py
@@ -23,12 +23,12 @@ from camel.messages.base import BaseMessage
 from camel.societies import RolePlaying
 from camel.tasks.task import Task, TaskState
 from camel.utils import print_text_animated
-from camel.workforce.prompts import (
+from camel.societies.workforce.prompts import (
     ROLEPLAY_PROCESS_TASK_PROMPT,
     ROLEPLAY_SUMMARIZE_PROMPT,
 )
-from camel.workforce.utils import TaskResult
-from camel.workforce.worker import Worker
+from camel.societies.workforce.utils import TaskResult
+from camel.societies.workforce.worker import Worker
 
 
 class RolePlayingWorker(Worker):
