+++ camel/societies/workforce/utils.py
