+++ camel/societies/workforce/worker.py
@@ -20,9 +20,9 @@ from typing import List
 from colorama import Fore
 
 from camel.tasks.task import Task, TaskState
-from camel.workforce.base import BaseNode
-from camel.workforce.task_channel import TaskChannel
-from camel.workforce.utils import check_if_running
+from camel.societies.workforce.base import BaseNode
+from camel.societies.workforce.task_channel import TaskChannel
+from camel.societies.workforce.utils import check_if_running
 
 logger = logging.getLogger(__name__)
