+++ camel/societies/workforce/workforce.py
@@ -28,21 +28,21 @@ from camel.models import ModelFactory
 from camel.tasks.task import Task, TaskState
 from camel.toolkits import SEARCH_FUNCS, WEATHER_FUNCS, GoogleMapsToolkit
 from camel.types import ModelPlatformType, ModelType
-from camel.workforce.base import BaseNode
-from camel.workforce.prompts import (
+from camel.societies.workforce.base import BaseNode
+from camel.societies.workforce.prompts import (
     ASSIGN_TASK_PROMPT,
     CREATE_NODE_PROMPT,
     WF_TASK_DECOMPOSE_PROMPT,
 )
-from camel.workforce.role_playing_worker import RolePlayingWorker
-from camel.workforce.single_agent_worker import SingleAgentWorker
-from camel.workforce.task_channel import TaskChannel
-from camel.workforce.utils import (
+from camel.societies.workforce.role_playing_worker import RolePlayingWorker
+from camel.societies.workforce.single_agent_worker import SingleAgentWorker
+from camel.societies.workforce.task_channel import TaskChannel
+from camel.societies.workforce.utils import (
     TaskAssignResult,
     WorkerConf,
     check_if_running,
 )
-from camel.workforce.worker import Worker
+from camel.societies.workforce.worker import Worker
 
 logger = logging.getLogger(__name__)
