+++ test/workforce/test_workforce_single_agent.py
@@ -16,7 +16,7 @@ import pytest
 from camel.agents.chat_agent import ChatAgent
 from camel.messages.base import BaseMessage
 from camel.tasks.task import Task
-from camel.workforce.single_agent_worker import SingleAgentWorker
+from camel.societies.workforce.single_agent_worker import SingleAgentWorker
 
 
 @pytest.mark.asyncio
